% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesClassify.R
\name{accuracy}
\alias{accuracy}
\title{Computes the accuracy from a confusion matrix}
\usage{
accuracy(confusionMat)
}
\arguments{
\item{confusionMat}{A confusion matrix}
}
\value{
The accuracy of a prediction given by
#-of-correct-predictions/total-#-of-predictions
}
\description{
Computes the accuracy from a confusion matrix
}
\examples{
#Yeah
}

