% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{accuracyScore}
\alias{accuracyScore}
\title{accuracyScore}
\usage{
accuracyScore(predLabels, trueLabels)
}
\arguments{
\item{predLabels}{predicted labels}

\item{trueLabels}{true labels}
}
\value{
Accuracy score value
}
\description{
Calculates accuracy for the given set of predicted labels and ground truth labels.
}
