% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesClassify.R
\name{bayesClassify}
\alias{bayesClassify}
\title{Take a list of texts and classify them using naive Bayes}
\usage{
bayesClassify(testSet, postList)
}
\arguments{
\item{testSet}{A list of character vectors. (Output from cleanData())}

\item{postList}{A list of two elements. The first is a data frame containing
likelihoods (and evidence) and the second is a vector of prior probabilities.
(This list is obtained from bayesProb().)}
}
\value{
A factor vector a prediction of either "cs", "math", "physics",
or "stat" for each text provided in the testSet list.
}
\description{
Take a list of texts and try to classify them using naive bayes. The sum of
the log-likelihoods of each word in each text in the list is computed
for each category. The log-prior is added to each result. Each text
is predicted to belong to the category with the maximal sum. (Words within
the test texts that were not observed in training are ignored.)
}
\examples{
#bayesClassify(cleanTestIn, bayesProbList)
}

