% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bernoulliNaiveBayes.R
\name{bernoulliNaiveBayes}
\alias{bernoulliNaiveBayes}
\title{Bernoulli naive bayes classifier for bag of words representation.}
\usage{
bernoulliNaiveBayes(formula, data)
}
\arguments{
\item{formula}{Describes the name of the column, which will be used as a label.}

\item{data}{Dataframe that contains training data.}
}
\value{
An instance of bernoulliNaiveBayes class trained on the given data.
}
\description{
Bernoulli naive bayes classifier implementation. This function creates an instance of bernoulliNaiveBayes
class and calculates apriori class probabilities as well as conditional probabilities. Uses Laplace smoothing
to ensure, that each conditional probability is greater than 0.
}
