% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanData.R
\name{cleanInData}
\alias{cleanInData}
\title{Read, clean, and format text data}
\usage{
cleanInData(filename)
}
\arguments{
\item{fileName}{A string of the name of the .csv file to be read. If the
file is not in the working directory, the full path name must be specified.
The "id" column is discarded when the .csv file is read, so the input text
.csv file and the category .csv file must be matched by row beforehand.}
}
\value{
A list of character vectors. All vector entries are strings of only
 lowercase , alphabetic characters.
}
\description{
Read, clean, and format text data from a .csv file. The function treats
each row of the .csv file as a single string. The function splits each
string into a vector of "words." A "word" is any consecutive series of
alphabetic characters. All spaces and non-alphabetic characters are
treated as word-separators and are removed. All "words" are coerced to
lower case.
}
\examples{
#cleanTrainIn <- cleanData("train_in.csv")

}

