% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanData.R
\name{cleanOutData}
\alias{cleanOutData}
\title{Read and format the known categories for text data from a .csv file}
\usage{
cleanOutData(filename)
}
\arguments{
\item{filename}{A string of the name of the .csv file to be read. If the
file is not in the working directory, the full path name must be specified.
The "id" column is discarded when the .csv file is read, so the input text
.csv file and the category .csv file must be matched by row beforehand.}
}
\value{
A factor vector of category names. The  vector must be in the
same order as the The factor has levels c("math","cs","physics","stat",
"category") which correspond to numeric values c(1,2,3,4,5) (if coerced).
}
\description{
Read and format the known categories for text data from a .csv file
}
\examples{
# Blarge
}

