% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesClassify.R
\name{confusion}
\alias{confusion}
\title{Computes a confusion matrix for a set of predictions and their known true
values}
\usage{
confusion(predOut, realOut)
}
\arguments{
\item{predOut}{The prediction results as a vector. (Best if factor vector)}

\item{realOut}{The real values as a vector. (Best if factor vector)}
}
\value{
The confusion matrix
}
\description{
Computes a confusion matrix for a set of predictions and their known true
values
}
\examples{
# What
}

