% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossValidation.R
\name{crossValidation}
\alias{crossValidation}
\title{Perform k-fold cross-validation}
\usage{
crossValidation(func, N, k, trainIn, trainOut, verbose = TRUE,
  innerRes = FALSE, ...)
}
\arguments{
\item{func}{The function you'll be cross validating
(in this case, testFunc()).}

\item{N}{The size of the sample you want to use}

\item{k}{The number of folds}

\item{trainIn}{The training data over which you'll be cross validating.
(Output from cleanData())}

\item{trainOut}{The categories for the training data}

\item{verbose}{Should the accuracy be printed after each iteration of the
cross-validation procedure? (Defaults to TRUE so you won't your computer
has crashed.)}

\item{innerRes}{Should all k intermediate results for the individual cross-
validations be returned? By default only the summary of all k-folds is
returned because it's messy otherwise.}

\item{...}{Any additional arguments for func.}
}
\value{
By default a list with two elements is returned. The first is the
summary confusion matrix and the second is the summary accuracy.
If innerRes == TRUE, then a list with two "sub" lists is returned.
The first "sub" list contains the default return value. The second "sub"
list is a list of all the confusion matrices and accuracies for each
individual iteration of the k-folds. These results are stored in a list
for each

A list of two elements. The first is a list containing the results
from the confusion() function for every trial. The second is a list containing
the summary of the results in the first list.
}
\description{
Perform k-fold cross-validation
}
\examples{
# crossValidation(testFunc,5000,10,cleanTrainIn,cleanTrainOut,
# alpha = 0.01, diffuse = 0.9)
}

