% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{getCVIndex}
\alias{getCVIndex}
\title{Provides the indices for the training and validation sets for k-fold
cross-validation.}
\usage{
getCVIndex(N, k)
}
\arguments{
\item{N}{The size of the total sample.}

\item{k}{The number of folds.}
}
\value{
A list of k index-vectors denoting which elements of the data set
is to be used for training.
}
\description{
Provides the indices for the training and validation sets for k-fold
cross-validation.
}
\examples{
# What
}

