% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HouseVotes84}
\alias{HouseVotes84}
\title{HouseVotes dataset}
\format{a data frame with 232 rows and 17 columns.}
\source{
\url{http://www.ics.uci.edu/~mlearn/MLRepository.html}
}
\usage{
HouseVotes84
}
\description{
This data set includes votes for each of the U.S. House of Representatives Congressmen on the 16 key votes identified by the CQA.
}
\details{
These data have been taken from the UCI Repository Of Machine Learning Databases
and were converted to R format by Friedrich Leisch.
}
\keyword{datasets}
