\name{multinomial_naive_bayes}
\alias{multinomial_naive_bayes}

\title{Multinomial Naive Bayes Classifier}
\usage{
multinomial_naive_bayes(x, y, prior = NULL, laplace = 0.5, ...)
}
\arguments{
\item{x}{numeric matrix with integer predictors.}

\item{y}{class vector (character/factor/logical).}

\item{prior}{vector with prior probabilities of the classes. If unspecified, the class proportions for the training set are used. If present, the probabilities should be specified in the order of the factor levels.}

\item{laplace}{value used for Laplace smoothing (additive smoothing). Defaults to 0.5.}

\item{...}{not used.}

}

\value{
\code{multinomial_naive_bayes} returns an object of class "multinomial_naive_bayes" which is a list with following components:
    \item{data}{list with two components: \code{x} (matrix with predictors) and \code{y} (class variable).}
    \item{levels}{character vector with values of the class variable.}
    \item{laplace}{amount of Laplace smoothing (additive smoothing).}
    \item{params}{matrix with class conditional parameter estimates.}
    \item{prior}{numeric vector with prior probabilities.}
    \item{call}{the call that produced this object.}
}
\description{
\code{multinomial_naive_bayes} is used to fit the Multinomial Naive Bayes model.
}

\details{
This is a specialized version of the Naive Bayes classifier, where the features represent the frequencies with which events have been generated by a multinomial distribution.

The Multinomial Naive Bayes is not available through the \code{\link{naive_bayes}} function.
}

\references{
McCallum, Andrew; Nigam, Kamal (1998). A comparison of event models for Naive Bayes text classification (PDF). AAAI-98 workshop on learning for text categorization. 752. \url{http://www.cs.cmu.edu/~knigam/papers/multinomial-aaaiws98.pdf}
}

\examples{
### Simulate the data:
cols <- 10 ; rows <- 100
M <- matrix(sample(0:5, rows * cols,  TRUE), nrow = rows, ncol = cols)
y <- factor(sample(paste0("class", LETTERS[1:2]), rows, TRUE, prob = c(0.3,0.7)))
colnames(M) <- paste0("V", seq_len(ncol(M)))
laplace <- 1

### Train the Multinomial Naive Bayes
mnb <- multinomial_naive_bayes(x = M, y = y, laplace = laplace)
summary(mnb)

# Classification
head(predict(mnb, newdata = M, type = "class"))
head(mnb \%class\% M)

# Posterior probabilities
head(predict(mnb, newdata = M, type = "prob"))
head(mnb \%prob\% M)

# Parameter estimates
coef(mnb)
}
\author{
Michal Majka, \email{michalmajka@hotmail.com}
}
\seealso{
\code{\link{predict.multinomial_naive_bayes}}, \code{\link[naivebayes]{tables}}, \code{\link[naivebayes]{get_cond_dist}}, \code{\link[naivebayes]{\%class\%}}, \code{\link[naivebayes]{coef.multinomial_naive_bayes}}
}

