% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinomialNaiveBayes.R
\name{multinomialNaiveBayes}
\alias{multinomialNaiveBayes}
\title{Multinomial naive bayes classifier for bag of words representation.}
\usage{
multinomialNaiveBayes(formula, data)
}
\arguments{
\item{formula}{Describes the name of the column, which will be used as a label.}

\item{data}{Dataframe that contains training data.}
}
\value{
An instance of multinomialNaiveBayes class trained on the given data.
}
\description{
Multinomial naive bayes classifier implementation. This function creates an instance of multinomialNaiveBayes
class and calculates apriori class probabilities as well as conditional probabilities. Uses Laplace smoothing
to ensure, that each conditional probability is greater than 0.
}
