% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bernoulliNaiveBayes.R
\name{predict.bernoulliNaiveBayes}
\alias{predict.bernoulliNaiveBayes}
\title{predict.bernoulliNaiveBayes}
\usage{
\method{predict}{bernoulliNaiveBayes}(object, data)
}
\arguments{
\item{object}{An instance of bernoulliNaiveBayes class.}

\item{data}{Dataframe that contains data to be classified.}
}
\value{
Vector of labels for the given data.
}
\description{
Predicts labels for the given data using previously trained classifier.
}
