% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinomialNaiveBayes.R
\name{predict.multinomialNaiveBayes}
\alias{predict.multinomialNaiveBayes}
\title{predict.multinomialNaiveBayes}
\usage{
\method{predict}{multinomialNaiveBayes}(object, data)
}
\arguments{
\item{object}{An instance of multinomialNaiveBayes class.}

\item{data}{Dataframe that contains data to be classified.}
}
\value{
Vector of labels for the given data.
}
\description{
Predicts labels for the given data using previously trained classifier.
}
