% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bernoulliNaiveBayes.R
\name{print.bernoulliNaiveBayes}
\alias{print.bernoulliNaiveBayes}
\title{print.bernoulliNaiveBayes}
\usage{
\method{print}{bernoulliNaiveBayes}(object, ...)
}
\arguments{
\item{object}{An instance of bernoulliNaiveBayes class to be printed.}
}
\description{
Prints an instance of bernoulliNaiveBayes class. This includes printing apriori class probabilities
as well as printing conditional probabilities.
}
