% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{profileFunc}
\alias{profileFunc}
\title{Profile a function to determine how much processing time each segment of
code within the function takes.}
\usage{
profileFunc(func, ...)
}
\arguments{
\item{func}{A function to be profiled}

\item{...}{The necessary arguments to be passed to the function}
}
\value{
A list of two elements. The first element contains the output
of func. The second element is the profile of func.
}
\description{
Profile a function to determine how much processing time each segment of
code within the function takes.
}
\examples{
#profileFunc(runAll())
}

