% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runAll.R
\name{runCV}
\alias{runCV}
\title{Runs cross validation without the user having to know anything about the code}
\usage{
runCV()
}
\value{
See crossValidation()
}
\description{
Runs cross validation without the user having to know anything about the code
}
\examples{
# Bleh
}

