% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runAll.R
\name{runTest}
\alias{runTest}
\title{Performs the full naive bayes procedure without the user having to know
anything about the code.}
\usage{
runTest()
}
\value{
Returns the prediction (it also writes a file called
"test_out_runTest.csv" to the working directory.)
}
\description{
Performs the full naive bayes procedure without the user having to know
anything about the code.
}
\examples{
# Ah
}

