% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossValidation.R
\name{testFunc}
\alias{testFunc}
\title{Helper for crossValidation}
\usage{
testFunc(trainIn, trainOut, testIn, testOut, ...)
}
\arguments{
\item{trainIn}{NA}

\item{trainOut}{NA}

\item{testIn}{NA}

\item{testOut}{NA}

\item{...}{NA}
}
\value{
NA
}
\description{
This is a function called by the crossValidation() function to help it
do its thing. This function has some of the other functions hard-coded
into it, so it's not really useable.
}
\examples{
# What
}

