% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tweet1}
\alias{tweet1}
\title{Tweets dataset}
\format{a data frame with 100 rows and 2214 columns.}
\source{
\url{https://www.figure-eight.com/data-for-everyone/}
}
\usage{
tweet1
}
\description{
This data originally came from Crowdflower's Data for Everyone library
}
\details{
"A sentiment analysis job about the problems of each major U.S. airline. Twitter data was scraped from February of 2015 and contributors were asked to first classify positive, negative, and neutral tweets, followed by categorizing negative reasons (such as "late flight" or "rude service")."
The data provided here is an altered and cleaned version version of the original source.
}
\keyword{datasets}
