% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{vectorizeAbstracts}
\alias{vectorizeAbstracts}
\title{Vectorize a list of text vectors}
\usage{
vectorizeAbstracts(trainIn, trainOut)
}
\arguments{
\item{trainIn}{A list of character vectors. (Output from cleanData())}

\item{trainOut}{A character vector denoting the Category to which each
element of the trainIn list belongs.}
}
\value{
A list of length C of character vectors, where C is the number of
unique categories in the trainOut vector.
}
\description{
Take a list of individual text vectors and use a vector of categories to
get a list of text vectors concatenated by category.
}
\examples{
## The number of texts
length(cleanTrainIn)

## The number of categories
length(vectorizeAbstracts(cleanTrainIn,cleanTrainOut))
}

