% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanData.R
\name{writeResults}
\alias{writeResults}
\title{Writes the results of the predictions in the specified format.}
\usage{
writeResults(output, filename = "test_out.csv")
}
\arguments{
\item{output}{The pediction from bayesClassify()}

\item{filename}{The name of the .csv file to which to write the prediction}
}
\description{
Writes the results of the predictions in the specified format.
}
\examples{
# Meh
}

