% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naive_bayes.R
\name{naive_bayes}
\alias{naive_bayes}
\title{Naive Bayes model}
\usage{
naive_bayes(X, y)
}
\arguments{
\item{X}{data matrix which stores observations as rows and
features as columns}

\item{y}{column vector that stores each observation's class labels}
}
\value{

}
\description{
Creates a summary matrix of the means and standard devations
for the data matrix split by classes and features. It also
calcaulates the prior class probabilities.
}
\examples{

}
