% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naive_bayes.R
\name{predict.naive_bayes}
\alias{predict.naive_bayes}
\title{Predictions for Naive Bayes}
\usage{
\method{predict}{naive_bayes}(object, x_new, ...)
}
\arguments{
\item{object}{instance of naive_bayes class}

\item{x_new}{test data matrix with observations as rows and features as columns}
}
\value{
predicted class labels
}
\description{
Makes predictions based on the fitted naive_bayes model by using a
Gaussian assumption that each variable is normally ditributed.
}
\examples{

}
