% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinomialNaiveBayes.R
\name{print.multinomialNaiveBayes}
\alias{print.multinomialNaiveBayes}
\title{print.multinomialNaiveBayes}
\usage{
\method{print}{multinomialNaiveBayes}(x, ...)
}
\arguments{
\item{x}{An instance of multinomialNaiveBayes class to be printed.}

\item{...}{other arguments passed to methods}
}
\description{
Prints an instance of multinomialNaiveBayes class. This includes printing apriori class probabilities
as well as printing conditional probabilities.
}
